\name{colSets}
\alias{colSets}
\title{plot colors}
\usage{
  colSets(type)
}
\arguments{
  \item{type}{palette names from the lists: Accent, Dark2,
  Pastel1, Pastel2, Paired, Set1, Set2, Set3}
}
\description{
  plot colors
}
\examples{
colors= colSets(type="Accent")
colors= colSets(type="Paired")
colors= colSets(type="Set3")
}

