% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PubMedWordCloud.R
\name{PubMedWordCloud}
\alias{PubMedWordCloud}
\title{Make a Word Cloud using the pubmed abstracts arrived at from
a pubmed query}
\usage{
PubMedWordCloud(SearchTerm, min.freq = 3, max.words = 250)
}
\arguments{
\item{SearchTerm}{character (1) pubmed search term}

\item{min.freq}{integer(1) minimum word frequency for wordcloud}

\item{max.words}{integer(1) maximum number of words for wordcloud}
}
\value{
generates a wordcloud figure
}
\description{
Make a Word Cloud using the pubmed abstracts arrived at from
a pubmed query
}
\details{
This method uses functions from the wordcloud,
}

